+++ src/main/java/dk/itu/models/OsmElement.java
@@ -4,10 +4,10 @@ import java.io.Serializable;
 
 public abstract class OsmElement implements Serializable {
     public abstract long getId();
-    abstract float getMinLon();
-    abstract float getMaxLon();
-    abstract float getMinLat();
-    abstract float getMaxLat();
+    abstract double getMinLon();
+    abstract double getMaxLon();
+    abstract double getMinLat();
+    abstract double getMaxLat();
     public boolean equals(OsmElement obj) {
         return this.getId() == obj.getId();
     }
